+++ src/metrics.rs
@@ -26,6 +26,22 @@ pub struct BitmapFontMetrics {
     pub bottom: f32,
 }
 
+/// Decoration dimensions for [`mplusfonts`](../mplusfonts/index.html).
+#[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct DecorationDimensions {
+    /// The offset from the baseline to the top of the decoration, defined as either _-100/1000
+    /// em_-size in the case of [`underline`] or _312/1000 em_-size in the case of [`strikethrough`]
+    /// for both **M<sup>+</sup> 1/2** and **M<sup>+</sup> Code**.
+    ///
+    /// [`underline`]: ../mplusfonts/struct.BitmapFont.html#structfield.underline
+    /// [`strikethrough`]: ../mplusfonts/struct.BitmapFont.html#structfield.strikethrough
+    pub offset: f32,
+    /// The height of the decoration, defined as _50/1000 em_-size for both **M<sup>+</sup> 1/2**
+    /// and **M<sup>+</sup> Code**.
+    pub height: f32,
+}
+
 impl BitmapFontMetrics {
     /// Metrics of the invisible bitmap font.
     pub const NULL: Self = Self {
@@ -51,6 +67,26 @@ impl BitmapFontMetrics {
     /// Returns the line height in pixels.
     pub const fn line_height(&self) -> u32 {
         let height = self.y_offset(Baseline::Top) - self.y_offset(Baseline::Bottom);
+
+        height as u32
+    }
+}
+
+impl DecorationDimensions {
+    /// The dimensions of the invisible decoration.
+    pub const NULL: Self = Self {
+        offset: 0.0,
+        height: 0.0,
+    };
+
+    /// Returns the _y_-offset for a line primitive in pixels.
+    pub const fn y_offset(&self) -> i32 {
+        self.offset as i32 + (self.offset % 1.0 > 0.0) as i32
+    }
+
+    /// Returns the stroke width for a line primitive in pixels.
+    pub const fn stroke_width(&self) -> u32 {
+        let height = self.height as i32 + (self.height % 1.0 > 0.0) as i32;
 
         height as u32
     }